queue_job: extract controller's runjob as JobExecutor
