+++ queue_job/tests/__init__.py
@@ -1,10 +1,12 @@
-from . import test_run_rob_controller
-from . import test_runner_channels
-from . import test_runner_runner
-from . import test_delayable
-from . import test_delayable_split
-from . import test_json_field
-from . import test_model_job_channel
-from . import test_model_job_function
-from . import test_queue_job_protected_write
-from . import test_wizards
+from . import (
+    test_delayable,
+    test_delayable_split,
+    test_executor,
+    test_json_field,
+    test_model_job_channel,
+    test_model_job_function,
+    test_queue_job_protected_write,
+    test_runner_channels,
+    test_runner_runner,
+    test_wizards,
+)
